add: 64-bit timebase + T_air_ref calibration constant (#2)
TIMER1's 16 MHz counter is software-extended to 64 bits (the raw 32-bit counter
wraps at 268 s — shorter than a run). BEACON/EVENT timestamps and the sync
offset are now 64-bit (DESIGN §2.6); the master still reports 32-bit ms
(tick/16000, 49-day wrap, matching the legacy HAL_GetTick) plus the 16 MHz $T
companion line.

T_air_ref documented as a bias-only constant (cancels for sensor-to-sensor
splits; scope-calibrated; moves to the per-set ID_SETUP table later).

Verified: $S ms + $T 16 MHz still correct at the 3 s sim spacing.

+++ traffic/device/firmware/src/protocol.h
@@ -1,9 +1,7 @@
 /* LoRa packet formats + time-sync helpers (DESIGN.md §2.5, §2.6).
  *
- * Timestamps are 16 MHz TIMER1 ticks. DESIGN specifies 64-bit fields with a
- * software 32->64 extension; bring-up uses the raw 32-bit tick (wraps ~268 s),
- * which is fine for the per-beacon offset math (done mod 2^32). 64-bit
- * extension is a later step.
+ * Timestamps are 64-bit 16 MHz TIMER1 ticks (DESIGN §2.6), software-extended
+ * from the 32-bit counter so they never wrap during a run.
  */
 #ifndef PROTOCOL_H
 #define PROTOCOL_H
@@ -23,7 +21,7 @@ typedef struct __attribute__((packed)) {
     uint8_t  type;       /* PKT_TYPE_BEACON */
     uint8_t  set_id;
     uint8_t  seq;        /* beacon sequence (wraps at 256) */
-    uint32_t m_tx_prev;  /* master TxDone tick of beacon (seq-1) */
+    uint64_t m_tx_prev;  /* master TxDone tick of beacon (seq-1) */
     uint16_t period_ms;
     uint16_t crc;        /* CRC16-CCITT over the preceding bytes */
 } beacon_t;
@@ -34,7 +32,7 @@ typedef struct __attribute__((packed)) {
     uint8_t  set_id;
     uint8_t  node_id;
     uint16_t ev_seq;
-    uint32_t ev_master_t;
+    uint64_t ev_master_t;
     uint8_t  flags;
     uint16_t crc;
 } event_t;
